feat(contact-center): modified-tests-and-renamed
