+++ packages/@webex/contact-center/src/metrics/behavioral-events.ts
@@ -330,7 +330,7 @@ const eventTaxonomyMap: Record<string, BehavioralEventTaxonomy> = {
     verb: 'fail',
   },
 
-  // EntryPoint API Events
+  // IEntryPoint API Events
   [METRIC_EVENT_NAMES.ENTRYPOINT_FETCH_SUCCESS]: {
     product,
     agent: 'user',
