+++ packages/@webex/contact-center/test/unit/spec/services/EntryPoint.ts
@@ -1,4 +1,4 @@
-import AddressBookAPI from '../../../../src/services/AddressBookAPI';
+import EntryPoint from '../../../../src/services/EntryPoint';
 import {HTTP_METHODS, WebexSDK, IHttpResponse} from '../../../../src/types';
 import {METRIC_EVENT_NAMES} from '../../../../src/metrics/constants';
 import WebexRequest from '../../../../src/services/core/WebexRequest';
@@ -8,17 +8,15 @@ import LoggerProxy from '../../../../src/logger-proxy';
 jest.mock('../../../../src/metrics/MetricsManager');
 jest.mock('../../../../src/logger-proxy');
 
-describe('AddressBookAPI', () => {
-  let addressBookAPI: AddressBookAPI;
+describe('EntryPoint', () => {
+  let entryPointAPI: EntryPoint;
   let mockWebex: WebexSDK;
   let mockMetricsManager: jest.Mocked<MetricsManager>;
 
-  const mockGetAddressBookId = jest.fn().mockReturnValue('test-address-book-id');
-
   beforeEach(() => {
     jest.clearAllMocks();
-    (WebexRequest as any).instance = undefined;
 
+    (WebexRequest as any).instance = undefined;
     mockWebex = {
       credentials: {
         getOrgId: jest.fn().mockReturnValue('test-org-id'),
@@ -41,7 +39,7 @@ describe('AddressBookAPI', () => {
     } as unknown as jest.Mocked<MetricsManager>;
     (MetricsManager.getInstance as jest.Mock).mockReturnValue(mockMetricsManager);
 
-    addressBookAPI = new AddressBookAPI(mockWebex, mockGetAddressBookId);
+    entryPointAPI = new EntryPoint(mockWebex);
   });
 
   describe('constructor', () => {
@@ -51,29 +49,31 @@ describe('AddressBookAPI', () => {
     });
   });
 
-  describe('getEntries', () => {
-    const mockEntries = [
+  describe('getEntryPoints', () => {
+    const mockEntryPoints = [
       {
         id: 'entry1',
-        name: 'John Doe',
-        number: '+1234567890',
-        organizationId: 'test-org-id',
+        name: 'Test Entry Point 1',
+        type: 'voice',
+        isActive: true,
+        orgId: 'test-org-id',
       },
       {
         id: 'entry2',
-        name: 'Jane Smith',
-        number: '+0987654321',
-        organizationId: 'test-org-id',
+        name: 'Test Entry Point 2',
+        type: 'chat',
+        isActive: true,
+        orgId: 'test-org-id',
       },
     ];
 
     const mockResponse: IHttpResponse = {
       statusCode: 200,
       method: 'GET',
-      url: '/organization/test-org-id/v2/address-book/test-address-book-id/entry',
+      url: '/organization/test-org-id/v2/entry-point',
       headers: {} as any,
       body: {
-        data: mockEntries,
+        data: mockEntryPoints,
         meta: {
           page: 0,
           pageSize: 100,
@@ -92,165 +92,159 @@ describe('AddressBookAPI', () => {
       jest.restoreAllMocks();
     });
 
-    it('should fetch address book entries successfully with default parameters', async () => {
+    it('should fetch entry points successfully with default parameters', async () => {
       (mockWebex.request as jest.Mock).mockResolvedValue(mockResponse);
 
-      const result = await addressBookAPI.getEntries();
+      const result = await entryPointAPI.getEntryPoints();
 
       expect(mockWebex.request).toHaveBeenCalledWith({
         service: 'wcc-api-gateway',
-        resource: '/organization/test-org-id/v2/address-book/test-address-book-id/entry?page=0&pageSize=100',
+        resource: '/organization/test-org-id/v2/entry-point?page=0&pageSize=100&sortOrder=asc',
         method: HTTP_METHODS.GET,
       });
 
       expect(result).toEqual(mockResponse.body);
-      expect(mockMetricsManager.timeEvent).toHaveBeenCalledWith(METRIC_EVENT_NAMES.ADDRESSBOOK_FETCH_SUCCESS);
+      expect(mockMetricsManager.timeEvent).toHaveBeenCalledWith(METRIC_EVENT_NAMES.ENTRYPOINT_FETCH_SUCCESS);
       expect(mockMetricsManager.trackEvent).toHaveBeenCalledWith(
-        METRIC_EVENT_NAMES.ADDRESSBOOK_FETCH_SUCCESS,
+        METRIC_EVENT_NAMES.ENTRYPOINT_FETCH_SUCCESS,
         {
           orgId: 'test-org-id',
-          bookId: 'test-address-book-id',
           statusCode: 200,
           recordCount: 2,
           totalRecords: 2,
           isSearchRequest: false,
           isFirstPage: true,
         },
-        ['behavioral', 'operational']
+        ['behavioral']
       );
-      expect(LoggerProxy.info).toHaveBeenCalled();
-      expect(LoggerProxy.log).toHaveBeenCalled();
+      expect(LoggerProxy.info).toHaveBeenCalledWith(
+        'Fetching entry points - orgId: test-org-id, page: 0, pageSize: 100, isSearchRequest: false',
+        {module: 'EntryPoint', method: 'getEntryPoints'}
+      );
+      expect(LoggerProxy.log).toHaveBeenCalledWith(
+        `Making API request to fetch entry points - resource: /organization/test-org-id/v2/entry-point?page=0&pageSize=100&sortOrder=asc, service: wcc-api-gateway`,
+        {module: 'EntryPoint', method: 'getEntryPoints'}
+      );
     });
 
-    it('should fetch entries with custom parameters', async () => {
+    it('should fetch entry points with custom parameters', async () => {
       (mockWebex.request as jest.Mock).mockResolvedValue(mockResponse);
 
       const params = {
-        addressBookId: 'custom-book-id',
         page: 1,
         pageSize: 25,
-        search: 'john',
-        filter: 'name=="John Doe"',
-        attributes: 'id,name,number',
+        search: 'test',
+        filter: 'type=="voice"',
+        attributes: 'id,name',
+        sortBy: 'name',
+        sortOrder: 'desc' as const,
       };
 
-      const result = await addressBookAPI.getEntries(params);
+      await entryPointAPI.getEntryPoints(params);
 
       expect(mockWebex.request).toHaveBeenCalledWith({
         service: 'wcc-api-gateway',
-        resource: '/organization/test-org-id/v2/address-book/custom-book-id/entry?page=1&pageSize=25&filter=name%3D%3D%22John+Doe%22&attributes=id%2Cname%2Cnumber&search=john',
+        resource: '/organization/test-org-id/v2/entry-point?page=1&pageSize=25&sortOrder=desc&search=test&filter=type%3D%3D%22voice%22&attributes=id%2Cname&sortBy=name',
         method: HTTP_METHODS.GET,
       });
 
-      expect(result).toEqual(mockResponse.body);
       expect(mockMetricsManager.trackEvent).toHaveBeenCalledWith(
-        METRIC_EVENT_NAMES.ADDRESSBOOK_FETCH_SUCCESS,
+        METRIC_EVENT_NAMES.ENTRYPOINT_FETCH_SUCCESS,
         {
           orgId: 'test-org-id',
-          bookId: 'custom-book-id',
           statusCode: 200,
           recordCount: 2,
           totalRecords: 2,
           isSearchRequest: true,
           isFirstPage: false,
         },
-        ['behavioral', 'operational']
+        ['behavioral']
       );
     });
 
     it('should handle API errors and track metrics', async () => {
       (mockWebex.request as jest.Mock).mockRejectedValue(new Error('Internal Server Error'));
 
-      await expect(addressBookAPI.getEntries()).rejects.toThrow('Internal Server Error');
+      await expect(entryPointAPI.getEntryPoints()).rejects.toThrow('Internal Server Error');
 
       expect(mockMetricsManager.trackEvent).toHaveBeenCalledWith(
-        METRIC_EVENT_NAMES.ADDRESSBOOK_FETCH_FAILED,
-        expect.objectContaining({
+        METRIC_EVENT_NAMES.ENTRYPOINT_FETCH_FAILED,
+        {
           orgId: 'test-org-id',
-          bookId: 'test-address-book-id',
           error: 'Internal Server Error',
           isSearchRequest: false,
           page: 0,
           pageSize: 100,
-        }),
-        ['behavioral', 'operational']
+        },
+        ['behavioral']
       );
       expect(LoggerProxy.error).toHaveBeenCalled();
     });
 
-    
+
 
     it('should not track metrics for subsequent pages in simple pagination', async () => {
-      const mockResponsePage2: IHttpResponse = {
-        statusCode: 200,
-        method: 'GET',
-        url: '/organization/test-org-id/v2/address-book/test-address-book-id/entry',
-        headers: {} as any,
-        body: {
-          data: mockEntries,
-          meta: {
-            page: 2,
-            pageSize: 100,
-            totalPages: 3,
-            totalRecords: 2,
-            orgid: 'test-org-id',
-          },
-        },
-      };
-
-      (mockWebex.request as jest.Mock).mockResolvedValueOnce(mockResponsePage2);
-
-      const result = await addressBookAPI.getEntries({page: 2});
-
-      expect(result).toEqual(mockResponsePage2.body);
-      expect(mockMetricsManager.trackEvent).not.toHaveBeenCalledWith(
-        METRIC_EVENT_NAMES.ADDRESSBOOK_FETCH_SUCCESS,
-        expect.any(Object),
-        expect.any(Array)
-      );
-    });
-
-    it('should track metrics for search requests on a valid non-first page', async () => {
       (mockWebex.request as jest.Mock).mockResolvedValue(mockResponse);
 
-      const result = await addressBookAPI.getEntries({page: 1, search: 'test'});
+      const result = await entryPointAPI.getEntryPoints({page: 2});
       expect(result).toEqual(mockResponse.body);
 
+      expect(mockMetricsManager.trackEvent).not.toHaveBeenCalled();
+    });
+
+    it('should track metrics for search requests on any page', async () => {
+      (mockWebex.request as jest.Mock).mockResolvedValue(mockResponse);
+
+      const result2 = await entryPointAPI.getEntryPoints({page: 2, search: 'test'});
+      expect(result2).toEqual(mockResponse.body);
+
       expect(mockMetricsManager.trackEvent).toHaveBeenCalledWith(
-        METRIC_EVENT_NAMES.ADDRESSBOOK_FETCH_SUCCESS,
+        METRIC_EVENT_NAMES.ENTRYPOINT_FETCH_SUCCESS,
         {
           orgId: 'test-org-id',
-          bookId: 'test-address-book-id',
           statusCode: 200,
           recordCount: 2,
           totalRecords: 2,
           isSearchRequest: true,
           isFirstPage: false,
         },
-        ['behavioral', 'operational']
+        ['behavioral']
       );
     });
 
-    it('should return cached data for repeat simple pagination calls', async () => {
+    it('should call API when requested page is not cached (cache miss)', async () => {
       (mockWebex.request as jest.Mock).mockResolvedValueOnce(mockResponse);
 
-      const first = await addressBookAPI.getEntries({page: 0});
-      expect(first).toEqual(mockResponse.body);
+      await entryPointAPI.getEntryPoints({page: 0});
+
+      (mockWebex.request as jest.Mock).mockResolvedValueOnce({
+        ...mockResponse,
+        body: {
+          data: mockEntryPoints,
+          meta: {
+            page: 1,
+            pageSize: 100,
+            totalPages: 2,
+            totalRecords: 2,
+            orgid: 'test-org-id',
+          },
+        },
+      });
 
       const callsBefore = (mockWebex.request as jest.Mock).mock.calls.length;
-      const second = await addressBookAPI.getEntries({page: 0});
+      (LoggerProxy.log as jest.Mock).mockClear();
+
+      const result = await entryPointAPI.getEntryPoints({page: 1});
       const callsAfter = (mockWebex.request as jest.Mock).mock.calls.length;
 
-      expect(second.data).toEqual(mockResponse.body.data);
-      expect(second.meta).toEqual(
-        expect.objectContaining({
-          page: 0,
-          pageSize: 100,
-          totalPages: 1,
-          totalRecords: 2,
-        })
-      );
-      expect(callsAfter).toBe(callsBefore);
+      expect(callsAfter).toBe(callsBefore + 1);
+      expect(result.meta.page).toBe(1);
+      expect(mockWebex.request).toHaveBeenCalledWith({
+        service: 'wcc-api-gateway',
+        resource: '/organization/test-org-id/v2/entry-point?page=1&pageSize=100&sortOrder=asc',
+        method: HTTP_METHODS.GET,
+      });
+      expect(LoggerProxy.log).toHaveBeenCalled();
     });
   });
 });
